requirements.txt
# PySpark 3.5.x + Delta 3.3.x — usar Java 17 (ver README)
pyspark>=3.5.0,<3.6.0
delta-spark>=3.3.0,<3.4.0

pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0

# Airflow 2.x (no 3.x: breaking changes vs DAGs actuales)
apache-airflow>=2.11.0,<3.0.0

streamlit>=1.50.0,<2.0.0
plotly>=6.0.0,<7.0.0

pytest>=8.3.0,<10.0.0
pytest-cov>=6.0.0,<8.0.0
chispa>=0.11.0,<0.12.0

black>=25.1.0,<26.0.0
isort>=6.0.0,<7.0.0
flake8>=7.1.0,<8.0.0
mypy>=1.14.0,<2.0.0
pre-commit>=4.0.0,<5.0.0

loguru>=0.7.3,<0.8.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<3.0.0
pyarrow>=18.0.0,<21.0.0
openpyxl>=3.1.0,<4.0.0
boto3>=1.35.0,<2.0.0
